Share FaaS target types, dependencies inference
