+++ src/python/pants/backend/google_cloud_function/python/rules.py
@@ -7,15 +7,17 @@ import logging
 from dataclasses import dataclass
 
 from pants.backend.google_cloud_function.python.target_types import (
-    PythonGoogleCloudFunctionCompletePlatforms,
     PythonGoogleCloudFunctionHandlerField,
     PythonGoogleCloudFunctionRuntime,
     PythonGoogleCloudFunctionType,
-    ResolvedPythonGoogleHandler,
-    ResolvePythonGoogleHandlerRequest,
 )
 from pants.backend.python.subsystems.lambdex import Lambdex
 from pants.backend.python.util_rules import pex_from_targets
+from pants.backend.python.util_rules.faas import (
+    PythonFaaSCompletePlatforms,
+    ResolvedPythonFaaSHandler,
+    ResolvePythonFaaSHandlerRequest,
+)
 from pants.backend.python.util_rules.pex import (
     CompletePlatforms,
     Pex,
@@ -55,7 +57,7 @@ class PythonGoogleCloudFunctionFieldSet(PackageFieldSet):
 
     handler: PythonGoogleCloudFunctionHandlerField
     runtime: PythonGoogleCloudFunctionRuntime
-    complete_platforms: PythonGoogleCloudFunctionCompletePlatforms
+    complete_platforms: PythonFaaSCompletePlatforms
     type: PythonGoogleCloudFunctionType
     output_path: OutputPathField
     environment: EnvironmentField
@@ -63,7 +65,7 @@ class PythonGoogleCloudFunctionFieldSet(PackageFieldSet):
 
 @rule
 async def digest_complete_platforms(
-    complete_platforms: PythonGoogleCloudFunctionCompletePlatforms,
+    complete_platforms: PythonFaaSCompletePlatforms,
 ) -> CompletePlatforms:
     return await Get(
         CompletePlatforms, UnparsedAddressInputs, complete_platforms.to_unparsed_address_inputs()
@@ -114,7 +116,7 @@ async def package_python_google_cloud_function(
     )
 
     complete_platforms = await Get(
-        CompletePlatforms, PythonGoogleCloudFunctionCompletePlatforms, field_set.complete_platforms
+        CompletePlatforms, PythonFaaSCompletePlatforms, field_set.complete_platforms
     )
 
     pex_request = PexFromTargetsRequest(
@@ -131,7 +133,7 @@ async def package_python_google_cloud_function(
     lambdex_pex, pex_result, handler, transitive_targets = await MultiGet(
         Get(VenvPex, PexRequest, lambdex_request),
         Get(Pex, PexFromTargetsRequest, pex_request),
-        Get(ResolvedPythonGoogleHandler, ResolvePythonGoogleHandlerRequest(field_set.handler)),
+        Get(ResolvedPythonFaaSHandler, ResolvePythonFaaSHandlerRequest(field_set.handler)),
         Get(TransitiveTargets, TransitiveTargetsRequest([field_set.address])),
     )
